requirements.txt
networkx==2.6.3
numpy==1.21.2
PyYAML==5.4.1
scipy==1.7.1